requirements.txt
requests
python-telegram-bot==13.7
python-dotenv